Module rdsync::cache

Caching module

Modules

cache_db

cache_table

Structs

CACHE

Global cache instance

CacheKey 🔒

Cache lookup key

Functions

add

Adds a new row to the cache table and the file database.

clear

Clears all entries from the cache.

delete

Deletes a row from the cache and the file database.

delete_table

Deletes a table and its associated entries from both the cache and the file database.

from_cache_string 🔒

Converts a cache key string into separate components (database, table, and key).

get

Retrieves a row from the cache table or the file database if not present in the cache.

keys

Retrieves a vector of all keys currently present in the cache.

to_cache_string 🔒

Converts the database name, table name, and key into a formatted cache key string.