Struct rdsync::cache::cache_table::Cache

pub struct Cache {
    pub data: HashMap<String, Row>,
    pub time_data: HashMap<String, TimeCache>,
    pub max_data_size: usize,
    pub current_data_size: usize,
    latest_time: u128,
    newest_time: u128,
}

Represents a cache storing rows and associated time metadata.

Fields
data: HashMap<String, Row>

HashMap storing rows with cache keys as the identifier.

time_data: HashMap<String, TimeCache>

HashMap storing time-related metadata for cache entries.

max_data_size: usize

Maximum size limit for the total data size in the cache (in bytes).

current_data_size: usize

Current size of the cached data.

latest_time: u128

Timestamp of the latest access in microseconds since the Unix epoch.

newest_time: u128

Timestamp of the newest access in microseconds since the Unix epoch.

Implementations

impl Cache

pub fn new(size_limit_mb: usize) -> Self

Creates a new cache with a specified size limit in megabytes.

pub fn get(&self, key: &str) -> Option<&Row>

Retrieves a row from the cache based on the provided cache key.

pub fn insert(&mut self, key: String, event: Row)

Inserts a new entry into the cache with the provided cache key and row.

pub fn keys(&self) -> Vec<String>

Retrieves a vector containing all cache keys.

pub fn clear(&mut self)

Clears all entries from the cache.

pub fn safe_time_insert(&mut self, key: &str, time_h: TimeCache)

Safely inserts time-related metadata into the cache, expanding if necessary.

pub fn update_last_accessed(&mut self, key: &str)

Updates the last accessed time for a cache entry based on the provided key.

pub fn delete(&mut self, key: &str)

Deletes a cache entry based on the provided key, reducing the current data size.